requirements.txt
Django<3.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
Pillow==7.0.0
pkg-resources==0.0.0
pytz==2019.3
django-polymorphic
django-polymorphic-tree
django-admin-sortable2
